[roll] Roll fuchsia [ffx][forward] Add machine-readable output support
This change adds support for machine-readable (`--machine json`) output
to `ffx forward`. This allows other tools to reliably parse the output
of the forwarding command, including the specific host and target ports
being used. This is especially useful when an ephemeral port (specified
as port `0`) is requested.

To support this, `ForwardSpec`, `ProtoSpec`, and `Direction` are now
serializable, and a new `ForwardMessage` enum is introduced for the
machine output schema.

Test: Manually ran `ffx forward --machine json tcp:0:22` and verified the
output JSON contained the dynamically allocated host port. Added new unit
and integration tests to verify the schema and the logic for resolving
ephemeral ports.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1441632
Original-Revision: 64e289a0f15e885bd383808b08021e71d6351a8a
GitOrigin-RevId: 30eb7acd546dbc5af4fdd9763398b2b5bb1a8bd7
Change-Id: I04b3f452aabbff78f9930b1f59af670d0666bbce