[roll] Roll fuchsia [ffx][target] Add target package gc
This adds ffx support to trigger a garbage collection on a target. It
is intended to be a drop-in replacement for running the command `fx
shell pkgctl gc`.
For use by ffx, the previous internal fidl fuchsia.space is renamed to
fuchsia.pkg.garbagecollector and offered to toolbox.

Original-Bug: 465722149
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1436016
Original-Revision: d9b5edf47a7fd14ca7b3425a34638dc881ed42cb
GitOrigin-RevId: 4d2b35dc5c3a8d6daca4a43ae77cfc7ad650e954
Change-Id: I4398534c9b7d2d8de1fd79172722824198cc10e8