[roll] Roll fuchsia [fdf][rust] Move some runtime APIs out of the public interface
Moves some calls that were in the public interface but with
'#[doc(hidden)]' out of the places where they're part of the public
api. These were made like this before `fdf_core` existed, but now
that it does we can properly restrict access to them.

Doing this because I intend to change the way shutdown observers
work to better support safe releasing of dispatcher references,
but want to get this move out of the way first.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1445757
Original-Revision: a44729dc1ebc74fb217527bc84e775633f37252b
GitOrigin-RevId: 390fa47723ade65e17727e25cf0730f8cf66a4c7
Change-Id: I6a68c1be7fcd463cf072e0113ef9eb5af4a0a73d