[roll] Roll fuchsia [ffx] Allow direct connections to re-materialize the Resolution
Previously, when we set up direct connection in the FHO environment, we
generated a resolution object and then used it for the whole life of the
environment. This is bad because Resolution fixes the IP address of the
device, so if we want to recover from an interruption that changes the
IP, we can't.

The fix is the usual pattern; add a wrapper around Resolution and let it
create a new resolution when needed.

Original-Fixed: 463689273
Test: Manually verified ffx log works when we bounce an emulator and its
      port changes.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1440568
Original-Revision: 28b5067c734f17407fe514aab7ccca26d1f37889
GitOrigin-RevId: d477ae895980058faff04f720e63c6d714fbb5c1
Change-Id: I4378040140b4508b967b526934909a6fd843aff3