[roll] Roll fuchsia [fidl/sockets] Improve FIDL definitions
- Rename "filter" to "matcher" everywhere.
- Update Control.DisconnectIp to match Diagnostics.IterateIp.
- Extract MatcherError and use consistently.
- Remove the `max_sockets` parameter from DisconnectIp.

That last one deserves some justification. While this check is useful in
theory, it requires either allocating a Vec of socket IDs, accepting
that the check is racy, or a relaxation of the check. For now, we're
just going to remove it since it's not a breaking change to add it (or
something similar) back in the future.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1445948
Original-Revision: 7148303fd1e261f4047bfddd819831d6703a300e
GitOrigin-RevId: ca32b6ddd4b661484b2e302c0e8fbdd79d7c7a78
Change-Id: I06f63a602afd7cea162d062e7a4e0e24c7b1d238