[roll] Roll fuchsia [netstack3] Implement Control.DisconnectIp for TCP
This piggybacks on the existing socket abort functionality, although it
sets a new Aborted error. It is also loosely based on the shutdown API
implementation, but with slightly different semantics since we're
bringing down the whole socket.

For better or worse, this change emulates Linux behavior in terms of
what operations continue to be valid on the SOCK (everything for UDP and
TCP listeners, but nothing for connected TCP sockets). Oddly, this
didn't require any extra work except for validation. In the future, we
may want to revisit this and allow resurrecting more kinds of sockets.

Original-Fixed: 459456549
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1445952
Original-Revision: f736c031ee6e71719eb3d5ca86239ba5e0430bf8
GitOrigin-RevId: 41f1e4b581e5f238e79f18ad0776c192784ad446
Change-Id: I3d58720d764b5b7799223c25fb77702ac2d12a8c