[roll] Roll fuchsia [elf_runner] Add allow_exception config for bad_handles policy
This adds a configuration to the elf runner's program block to enable
the bad handle policy with an action of ALLOW_EXCEPTION. This asks the
kernel to raise a non-fatal exception when a violation of the bad handle
policy is detected. Enabling this configuration will allow us to
identify violations of the bad handle policy without changing behavior
(outside of timing) to let us gradually fix issues before ramping up
enforcement of the bad handle policy.

The new syntax for bad handle job policy configuration is:

  program: {
    job_policy_bad_handles: "deny" | "allow_exception",
  }

This also adds docs for this field to the elf runner documentation.

Original-Bug: 42151766

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1445735
Original-Revision: 00d411a7c351ae5160e275cf26e532114eb76dd3
GitOrigin-RevId: 3b5dc184c795172164c997eab475407ef8d1a4f1
Change-Id: If903a2f318fb4a5053ea2c9a56ef0daf91405298