[roll] Roll fuchsia [assembly][vbmeta] Add footer support
Rather than being standalone, a VBMeta may be found appended to an image
or amongst partition contents. In this case - which will be the case for
pVM VBMeta - it is identified through a "VBMeta footer". We add support
for that here.

Original-Bug: 449201661
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1426935
Original-Revision: 3e8b908d205fec80b40d27078dce12f59c19e911
GitOrigin-RevId: ae554bed74a9ea6a245b3349937cbec99ba842e3
Change-Id: Icb660d961e2405d8676171c13befb9472c3b4eea