[roll] Roll fuchsia [wlan][test] Log test start/end in WifiBaseTest
This change adds a logline to the Fuchsia syslog at the setup and
teardown steps for tests that use WifiBaseTest, which helps correlate
syslogs with test logs. The same logging already exists for tests that
inherit FuchsiaBaseTest, so as tests are migrated out of the
//src/testing/end_to_end/antlion dir, they will still retain this
logging.

Original-Fixes: 465799197
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1446687
Original-Revision: d85ca2f4d7d8c761dcdbc1c0630cf176004b7083
GitOrigin-RevId: f04e5786bd9e3d1eda7390c55229c31068edf156
Change-Id: If41e5c0a70647f956d19b5540603a66606e165e3