[roll] Roll fuchsia [sestarnix] Add indices_of_set_bits
For most of the current users of ExtensibleBitmap.spans, what is
really wanted is the integers found within the spans representing the
indices of bits set in the ExtensibleBitmap, and a second iteration is
used at each .spans callsite to get these indices. Let's just offer the
iteration yielding the indices natively from ExtensibleBitmap itself.

Fun fact: I drafted this as "set_bit_indices", but then could not pass
self-review with a method with a name that started with "set_" but that
was not mutative. 😐

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1447088
Original-Revision: 7931606ea597d32732eeed606c1134a465212068
GitOrigin-RevId: 026bca069b75443250607ce9e69c1aece7f5aa51
Change-Id: Ie2379834fe6869fd6e480b3935f45a6509332227