[roll] Roll fuchsia [drivers] Add power entities to protocols
Add fields to pass power-related handles in the driver start
args. This will be a way to deliver the handles both to the
driver how, which directly receives the args, and then on to
the driver which gets the args forwarded to it by the driver
host.

Original-Bug: 454169735
Original-Fixed: 462180937

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1398853
Original-Revision: a0294efacb1ff82abc007a8a5bee3e8426e15d88
GitOrigin-RevId: 39233028c86913f291bba3b6d4999bffc25c5637
Change-Id: I3f72b3311b3412dfe1b3d5a096b2c6dd6d2eed0f