[roll] Roll fuchsia [usb_vsock_host] Remove cid argument from listen()
When I wrote this library I added the ability to listen for incoming
connections either from anywhere, or only from specific source
addresses. I misunderstood POSIX socket semantics and thought the
latter was a standard feature. It is not, and it's a source of some
complexity.

Ripping it out saves us some indirection, and there are no users of the
feature.

Test: unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1447112
Original-Revision: 6dc6aec9260a63803fa999d1406db61a47d08ed0
GitOrigin-RevId: 22550f2c405a3df5395a673504a9f097892c4371
Change-Id: I029ece4e9a3042dddadc60499a7555a0ed098d84