[roll] Roll fuchsia [ffx] Kill ssh.allow_fdomain
This config key lets users turn off connecting to the target with
FDomain. If a plugin uses FDomain there will always be at the least
local FDomain emulation, but setting this to false forces that
emulation to happen.

This key has a different default for normal and strict mode. In strict
mode it defaults to true, in normal mode it defaults to false. Since
infra has migrated to strict, all of our infra tests have been running
with this set to true.

This key feels a bit obscure, it already messed up one instance of me
asking a user for data because I forgot about it. Since it's set to
true in infra we've already tested pretty heavily. I'd like to just get
rid of this key and just always use FDomain if we have it. I think
that's less likely to cause user headaches than porting all the tools
then having a flag day where we flip this anyway. I'd rather discover
the issue incrementally as we port incrementally.  The only reason to
keep it around would be as a temporary mitigation for a user-discovered
issue, but I think the kind of issue where it would be useful is
unlikely.

Test: Existing tests should suffice
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1447113
Original-Revision: 151039b1655805ba65305de560599e748f523dbf
GitOrigin-RevId: f34447dfe846d35250481b265e0d89f0262eaad3
Change-Id: I4b33f03cd07fdc61e47f0e49140ebba55eef4f0a