[roll] Roll fuchsia [wlan] Create CoreBaseTestClass and use FuchsiaWlanBaseTest for core end-to-end tests
The largest part of this change is transitioning core end-to-end tests
off BaseTestClass to FuchsiaWlanBaseTest. The latter includes standard
functionality for Fuchsia end-to-end tests so that the classes defined
for WLAN core end-to-end tests don't need to provide that functionality
themselves.

The CoreBaseTestClass makes a connection to
fuchsia.wlan.device.service/DeviceMonitor and asserts exactly one
FuchsiaDevice is available.

Part of this change modifies ConnectTransactionEventHandler to take an
asyncio.AbstractEventLoop as an argument to step around the honeydew
async limitations. This will be unwound in a later change.

Also, a basic ping function is added ConnectionBaseTestClass for
FirmwarePowerModesTest. As other tests need this method, we may need to
move it elsewhere, even into honeydew as an affordance.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1441530
Original-Revision: 3efd3cdc20614022f4368aa7c5d9ad8f74adb6bb
GitOrigin-RevId: dbda3026f99e8f596ec4208bee266134888ab6fc
Change-Id: Ia8c051ade265486db2a117bf181162107fa3e729