[roll] Roll fuchsia [assembly][vbmeta] Introduce VBMeta style
This change introduces the notion of VBMeta "style" as a new VBMeta
board config option, encoding its 'shape' and location. The common case
is the current scheme for Fuchsia of standalone VBMeta, and this remains
the default for the new option. The other styles introduced are
variations on the theme of what is expected for Android protected VMs:

* "android-pvm-auto": Generates the VBMeta expected by Android protected
  VMs as a footer to the product bundle's QEMU kernel (i.e., boot shim),
  where the choice of ramdisk (i.e., ZBI) hash descriptor 'partition
  name' which controls a UART-related debug policy - is "initrd_normal"
  in user builds and "initrd_debug" in non-user builds.

* "android-pvm-normal": Similar to "android-pvm-auto", except that the
  ramdisk hash descriptor partition name is fixed as "initrd_normal".

* "android-pvm-debug" Similar to "android-pvm-auto", except that the
  ramdisk hash descriptor partition name is fixed as "initrd_debug".

Setting these for style will currently hit a todo!() and will be
implemented in a follow-up change.

One alternative name considered instead of "style" was "scheme", but
that seemed like it might ambiguously suggest a variation of the VBMeta
file format itself rather than a particular specialization of it.

Original-Bug: 449201661
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1425321
Original-Revision: 6428be989e162ec75089d974aba25cd2ce5f3fac
GitOrigin-RevId: 00b3424a200e3acaeeab1f848c44d6244c3d0a35
Change-Id: I08b5274f1545b47832420c64f57d657afc746791