[roll] Roll fuchsia [zircon][channel] Do not clean up handles after msg_put_handles
The zx_channel_write and zx_channel_write_etc syscall implementations
use the helper function channel_write. This helper sets up a deferred
cleanup lambda which closes all provided handles if an error is
encountered in the operation before the handle list is dealt with. This
moves cancelation of this cleanup action to before calling
msg_put_handles since msg_put_handles already deals with every handle in
the handle list according to the operation. In particular, if an
operation on one handle in the list fails then msg_put_handles takes
care to still perform the appropriate actions on other handles. The
cleanup closure doesn't know which handles have been dealt with by
msg_put_handles and it may attempt to close an already removed handle.

This also tightens up the error handling in msg_put_handles to always
close the source handle when the requested operation is not a duplicate
operation. This matches the system call documentation and function level
documentation.

Original-Fixed: 467142666
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1445128
Original-Revision: 6a32dca55b0650d51ca14d354badf816451bcd32
GitOrigin-RevId: 33b609e35f0830da05b2557deb66b75552581148
Change-Id: I1a4061deafa48a7e4506d43a67a127f3843f8d75