[roll] Roll fuchsia [security] Update Scrutiny structured config tests
This introduces a component_tree_config file to allow structured config
tests to use dynamic routing information. This is needed because the
structured_config tests create a test product bundle, which seem to
inherit a number of changes to core and is not well-insulated.

Pointing to the platform-level config might work, but this change will
opt to use a test-specific copy.

Original-Bug: b/468009955
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1446192
Original-Revision: 879aaf1c46529dc4bc3dcc4dbb8323658bbf6685
GitOrigin-RevId: 34c309736bde68e3a5a18ea295c8669c20008125
Change-Id: I6a3205818a576320d017ab0c5202cbd34f638dc2