[roll] Roll fuchsia [fuchsia-controller] Cancel pending tasks upon external cancellation
The FidlClient creates background tasks to wait on a handle being ready
to read, a staged message becoming available, or an epitaph on the
handle. Without structured concurrency, cancellation of the task that
created those tasks would result in the background tasks printing errors
that they were still pending when destroyed.

This change uses asyncio.TaskGroup to create the background tasks in a
manner that propagates the cancellation of the parent task to each
background task.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1446968
Original-Revision: 28466543a4e664808ef6c62faec63f9870b3e19f
GitOrigin-RevId: 5b28a988342929d11446dc8cdb5a5796b0529c6e
Change-Id: I6601018fc505a346c0e7f2045b4f5bfb45b0d160