[roll] Roll fuchsia [starnix][task] Delay releasing ThreadGroup
The `release` implementation for ThreadGroup is now just various
asserts. This CL moves those those asserts to happen during `drop` and
makes `release` into a noop.

Before this CL, we could crash when we had a delayed releaser that
touched `task.thread_group` because we nulled out that field as we
released the Task. Now, we don't null out that field, which means we
never enter the state in which we can crash.

A future CL will remove the Releaseable trait from ThreadGroup and just
use Arc<ThreadGroup> instead of OwnedRef<ThreadGroup>.

Original-Fixed: b/466174458
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1448395
Original-Revision: 52c7c42d96260f667aa760eddbc41e0220526ee1
GitOrigin-RevId: 98ed25e9f409286cf8a9b2eeae8d2ff65ae04742
Change-Id: I615612387c9c7d2459271efc26c97b31b31df567