[roll] Roll fuchsia [pkg] Use blob protocols in tests
Use the blob protocols to read and write blobs in the blobfs-ramdisk and
pkg-resolver tests.

All but 1 of the blobfs-ramdisk tests now run for both blobfs and
fxblob. The 1 exception is the concurrent creation test because blobfs
doesn't support it.

Original-Bug: b/337261333
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1433754
Original-Revision: 766b133cb2718a7e8c200746adf275754d758d11
GitOrigin-RevId: 6cd1c618cc3bed36df46b68eca70e8b6a9d4ec49
Change-Id: Ic21bab4e665b1a3233638ff99fd735fd42e8cf81