[roll] Roll fuchsia [components] DirConnector in fuchsia.component.sandbox to match internal
fuchsia.component.sandbox was created before much of the underlying
implementation had been migrated to the new runtime capabilities. During
this migration we discovered that some use cases called for being able
to pass flags and a subdir when opening a dir connector, despite the
preexisting API not supporting this.

A replacement for fuchsia.component.sandbox is in the works that will
address this along with other shortcomings, but some users are blocked
on the ability to use the full directory connector API.

To unblock these users in the short term, this commit reworks the
DirConnector APIs in fuchsia.component.sandbox to support passing these
arguments when opening the directory connector.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1390873
Original-Revision: 357f4d256072502b5c2968d668c21ce5b4f92bde
GitOrigin-RevId: 069204fec902f5826b1f0d7c92e3ca5ad4eaaa60
Change-Id: I9d97768c486bad4eb43f5da32d006fe1a99741ed