[roll] Roll fuchsia [sshd-host]: better handle logging of sshd exit code
When the service OnStop method is called, check the exit code of the
sshd child component, if it is 255 that's the signal that the client
terminated the connection.

Test: Invoke ffx log --component klog --component core/sshd-host. Then
invoke in another terminal `ffx -c ssh.controlmaster='none' log` and
wait until logs appear. Then CtrlC the log and note that the error is
not present at the WARN level for sshd-host.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1449298
Original-Revision: e3fd42ee991b8a7d1de55de8005c1514a0c1b7b3
GitOrigin-RevId: 0550420e2c0398d97dc06e67859505b17b4536d4
Change-Id: Ie5f8290af7e8b2a218a4bfe5d04f7f857049ef5e