[roll] Roll fuchsia [starnix] SysRq crashes Starnix rather than rebooting
When "c" is written to SysRq, panic the Starnix kernel directly
rather than rebooting the whole device. This is useful in
testing a recovery / reporting logic and avoiding the confusion
between the different failure modes.

The existing rebooting logic is moved to "b" which seems to be a better
fit.

Test: starnix_sysrq_integration_test

Original-Bug: 394953531
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1447888
Original-Revision: 6b61874ddb100e91b58b8dac2f06b1cb446536e3
GitOrigin-RevId: 8eaf430853a37e1c86c16521a5569da99fb9a21d
Change-Id: I4a4613a41ea4ae6eaa6001176c4b9e672776cddd