[roll] Roll fuchsia [rust][zx] Update tests to satisfy bad handle policy
This removes tests that interact with system calls only to the extent of
triggering the bad handle policy and configures the suite's bad handle
policy to raise exceptions.

If we want to exercise the syscall bindings for wrappers that interact
with privileged or otherwise specialized object types we could use the
fake-object library to interpose on the zx_.. symbols.

Original-Bug: 42151766

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1446196
Original-Revision: 94dd0151873c64ab6e8d07a9c9d5e687ab7e325e
GitOrigin-RevId: ce74289859317ccddaf8d777cccbf0163f52b2aa
Change-Id: If5e8b078edd302605cb64323e6da928904169c2a