[roll] Roll fuchsia [component_manager] Reduce number of items in realm_query::manifest_test to speed it up
This adds assertions on the size of the serialized manifest, and then
reduces the number of items it contains so that it fits in more than
one, but not more than 3, zircon channel messages.

This should significantly speed up this test on some platforms.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1449016
Original-Revision: 0cd0b6054ab98f9fb1adc68b3b718a017e45c375
GitOrigin-RevId: a51be37a9c0231a7945b260194174e56590fb9ae
Change-Id: I2fea7db23c2e5abd676ebdd86a145f88d74c9e66