[roll] Roll fuchsia [userboot][core-tests] Expose when midr.txt is emitted
Currently we only emit midr.txt when ZX_DEBUG_ASSERT_IMPLEMENTED is
nonzero. However, we are trying to run core tests in a configuration
where it is 0 (*-no_kernel_debug) and are failing to skip the
ProcessTest.GetInfoRuntime* tests in that context since there is no
midr.txt to make the "is Cavium?" check.

Instead of having the core test logic hardcode an expectation around
when a midr.txt is present, we introduce a constant in libuserabi that
can be consulted instead. When it indicates that no midr.txt is present
then we will always skip this test.

Original-Fixed: 467750183
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1447011
Original-Revision: fe30e3e8769e7f4246a320a4fb28d69360e7261f
GitOrigin-RevId: 79b08b283ca1cb1a7694a31dcbb09c320a9459ff
Change-Id: I607b73880a6392811c3a4009451474569cf27cb9