[roll] Roll fuchsia [bazel][fidl] Define an IDK atom when appropriate
Move the JSON summary file to the main build directory to comply with
symbolic macro naming requirements, and remove `gen_dir` from `fidlc()`
since it is unused.

Define //sdk/fidl, which contains atoms for FIDL libraries in the IDK.

Migrate fuchsia.hardware.spi.businfo and fuchsia.math to Bazel as
examples of unstable and stable, respectively, libraries in the IDK.

Always place `.fidl` files in the root of the library's directory in the
IDK to simplify logic. Add an assert to the GN template to ensure this
is not a change in behavior.

Original-Bug: 428285014
Test: fx bazel build --config=fuchsia //sdk/fidl
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1444484
Original-Revision: cf97137ec2d993d4aed2765cc0a56c6037d8ce47
GitOrigin-RevId: adeb81f2689fa6220ee91fb69ceb34ff3e471926
Change-Id: I9666b4b7c4ecf0595029def6c7b516c8f9ee0960