[roll] Roll fuchsia [owners] Remove gevalentino@google.com from zircon/kernel/lib/devicetree/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/gevalentino@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1452489
Original-Revision: f07a3d9a66d6c17c911e2a90892caed83c63ecdc
GitOrigin-RevId: 4b0251fb391dccad70c157abf82ceb97ea2365cb
Change-Id: I2c44bfd5bda91ad524c8d1b57cd84074baedbdbb