[roll] Roll fuchsia [owners] Remove gevalentino@google.com from zircon/kernel/lib/power-management/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/gevalentino@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1452513
Original-Revision: 437bfe08d463e02a018633d9a5a805e926478560
GitOrigin-RevId: af15f7dd5332dd7fd34b4913147a5a59742960be
Change-Id: I76c1fd81733258e8f5abcc5a228269684f3deddf