[roll] Roll fuchsia [owners] Remove gevalentino@google.com from zircon/system/ulib/uart/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/gevalentino@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1452490
Original-Revision: eb34aa585705a372ddc605199d0e184d2ff6080c
GitOrigin-RevId: 061e2a525e3667e608a0421815928412cbab6bcf
Change-Id: I3bab7fc5b83a624b4eb17824190c4409398cd794