[roll] Roll fuchsia [owners] Remove gevalentino@google.com from zircon/system/utest/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/gevalentino@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1452514
Original-Revision: 6bb1559a3c5e3d45590a47c8d5cb91c2514f4efa
GitOrigin-RevId: f0293f05ee080f23f9095a43c27ac20a58b6c192
Change-Id: I0b35338425bdf9ecb9f3c7480daa56ae90ba9714