[roll] Roll fuchsia [bt][gap] Commit device class to persisted BondingData
Add the device class field to the general BondingData struct defined in
the fuchsia-bluetooth crate. Also add an optional inspect node for this
property if it is set.

Update bt-gap to save and load this field so that it is persisted across
reboots.

Note that in practice, the device class will continue to not be present
in sys.Access/WatchPeers updates from the host stack as bt-host is not
updated to parse & save this field.

Original-Bug: 468338802
Test: fx test bt-gap-unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1448211
Original-Revision: 987fc902b75ad50820d85b2c75575b631db80547
GitOrigin-RevId: 6e7e81f5a1a6280ae57f9e99a13370ce493d48ec
Change-Id: Idb9fc4c99e5fd7a8073ed665722d5f0774d6c016