[roll] Roll fuchsia [owners] Remove gevalentino@google.com from src/lib/zxdump/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/gevalentino@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1452488
Original-Revision: 7f22f5e18ba7e54e74d15ea921c35c1b0eb88d61
GitOrigin-RevId: f676c1ac37e721f4506ef57e8098ab9ed0c9ef77
Change-Id: I09333cbb5b4ca07723749a35afbea24123407261