[roll] Roll fuchsia [owners] Remove gevalentino@google.com from zircon/kernel/lib/userabi/userboot/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/gevalentino@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1452510
Original-Revision: 857d99d09c2b77c1cf9bb098ee437c8105dff7be
GitOrigin-RevId: 62ffcb246ce262663fcf3e2811df6c281eda1ddd
Change-Id: Ia5a4e22b63c97330d35187de84934edf65362c43