[roll] Roll fuchsia [bt][avrcp] Rework connection logic to be on-demand
Previously, the AVRCP profile would automatically attempt to establish
control and browse connections as soon as a peer's services were
discovered. This is inefficient and has been reworked to an on-demand
model.

This change modifies the peer connection logic to only initiate AVCTP
control and browse connections when `GetControllerForTarget` or
`GetBrowseControllerForTarget` is explicitly called by a client.

`request_control_connection` and `request_browse_connection`
are exposed on the `RemotePeerHandle` to trigger connection attempts
from the FIDL service layer when a controller is requested.

Test: fx test bt-avrcp-tests
Original-Bug: 450144120

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1445968
Original-Revision: f8a4fcafecfbb7bb8771e6b667cbf01b5ee94a7b
GitOrigin-RevId: 29d097f157e1f912172645e445c20d57bbbedb42
Change-Id: I416bbc40b15fd937b1321b80393e42f70a1c60f6