[roll] Roll fuchsia [cog] Pass repository root from shell script to git.py
The repository root is now determined in the shell script and passed to git.py via a new --repository-root argument. This simplifies the Python script by removing the need to find the root itself. A check is added in git.py to ensure the provided repository root is within a cog workspace.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1453270
Original-Revision: 4f7f1b1621e4973deb0462467248170cff024af1
GitOrigin-RevId: 00da44290f1b321b4b6554d8481d8c00edc5e91d
Change-Id: I560583abea732002935cb6d72d41e98c7517ca5e