[roll] Roll fuchsia [fidl] Add create_proxy_and_stream to ZirconClient
ZirconClient is just a zero-sized struct that can be a receiver for
various utility methods that we'd normally call as receiverless
functions. This lets us have a common API between normal and FDomain
handles, which usually require a receiver as they refer to a
connection.

This patch just adds one more method to ZirconClient that was missing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1452815
Original-Revision: 1bd1a321e212e0f50644fc9756ff5ce48559e208
GitOrigin-RevId: ee21d361ce28d2e612e113f0b0f14d714f1e29aa
Change-Id: I4c5ef5e7d8ac21f461c668c94decc3dc939a9084