[roll] Roll fuchsia [kernel][vm] ReadWriteShrinkRace handle eviction
Spin up a page-fault handler thread to re-supply pages as they are
evicted. This avoids a possible race that we are not testing for, in
which a page could be supplied and then evicted before the read request
completes.

Original-Bug: 469169595
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1452968
Original-Revision: 578a2cc471a1c956bd6bd1c6ef2fe9bd3e12c4a1
GitOrigin-RevId: f064edea1fc0903fc7dda6077b3992d54e9deaf7
Change-Id: If23feb0928c7d2e89c78f4f7905bcf9b7b2b51c9