[roll] Roll fuchsia [botanist] Fix logic for when to mount shared_data dir.
This was only being created when running ffx monitor which only gets run
if we expect SSH. However, we were determining whether to mount the dir
in the nsjail invocation for host tests based on whether the env var was
set which we set if the experiment was enabled. This CL fixes the logic
to only set the env var when both expectsSSH is true and the ffx monitor
is enabled. This was previously causing failures in a boot test which
then immediately exited and tried to clean up the emulator before
finishing the task. However, it looks like ffx emu stop command didn't
actually kill the emulator (perhaps because it was run before the
emulator was fully set up) so it leaked the process to the next task,
causing failures like the attached bug. Thus, this CL also always kills
the emulator process just in case.

Original-Bug: 465528150
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1454208
Original-Revision: 8f8cac814927b623ff5b40841ab4f326c47bfa5c
GitOrigin-RevId: 34160402574cb15654c85ed97260957efcdcce3a
Change-Id: I30b143b0ad0b3e88b3e56016daaecdd0a7369eba