[roll] Roll fuchsia [starnix] Make Starnix kernel have a single_run durability
In practice this means that if the kernel ever gets stopped (notably
because the process associated with it died) it is automatically destroyed.
This simplifies cleanup: without this change in some cases the stopped
starnix_kernel component would linger in the component list.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1453688
Original-Revision: 4b42faa96bbbb5dd36e4dc24b6affe651441e0b8
GitOrigin-RevId: ab3e25b5d57394997a2ec0e381d3bf1f149c5f1d
Change-Id: Ia446d5856ca97bdf9941b55ce51bfe2d431f8bdd