[roll] Roll fuchsia [syslog] Support file = nullptr in LogMessage
The code in LogMessage::~LogMessage clearly attempts to handle the
case where `file_` is null, but the constructor calls
StripDots/StripPath on the string, which require a non-null
pointer. We now avoid calling these when the file argument is null.

Original-Bug: 456458551
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1454189
Original-Revision: c2db7323029a93d7992d3045fbfe9ff39689d441
GitOrigin-RevId: efd400eabf98abb3124b0e903912f052a02ff49e
Change-Id: I6bd85f085b6cba32ba6cf84a42e791f069e695ea