[roll] Roll fuchsia [feedback] Make report store size configurable
The report store persists all non-snapshot data if connectivity is
unavailable, including minidumps. For some processes, such as the
Starnix kernel, minidumps can be quite large and may exceed Feedback's
report store limit of 500 KiB for /cache.

This CL adds a flag to Feedback's config so that assembly can specify
the size of the report store. This CL does NOT change the current size
of the store. A follow-up CL will increase the store size for products
that specify the "large_disk" Feedback assembly flag as "true."

Original-Bug: 465548680
Test: `fx test //src/developer/forensics`
Test: `fx test
//src/lib/assembly/platform_configuration:platform_configuration_test`

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1444818
Original-Revision: 91905d554345130e22f3cb895878e204c48f8434
GitOrigin-RevId: f28e6ceec1ac00863c50f85feb469f8f0cb93960
Change-Id: I6024c92f77db98021c9663a1ddbb083a7d1378cc