[roll] Roll fuchsia [wlan] Use Constant for channel reporting
Reporting channel even when we are disconnected from a network makes the
visualization of this data confusing, since the other data (phy rate,
RSSI) interpolates 0 when we are not actively connected to a network.
Using a constant fill for channel data means that we will only see data
reported when we are connected to a network.

Original-Fixed: 466071471
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1441519
Original-Revision: 04748b9ea1f760260a1ae302788104bf3803f78b
GitOrigin-RevId: bbdf9f8ddbba6027f94494e2b0bac39083879234
Change-Id: I9d084454998a9154e52046a545a14c38507ccf7e