[roll] Roll fuchsia [core-tests] More gracefully handle absent midr.txt
Currently we only emit midr.txt for arm64 when
ZX_DEBUG_ASSERT_IMPLEMENTED in the kernel is nonzero. However, we are
trying to run core tests in a configuration where it is zero
(*-no_kernel_debug) and are failing to skip the
ProcessTest.GetInfoRuntime* tests in that context since there is no
midr.txt to make the "is Cavium?" check.

In this change we update the "is Cavium?" logic to allow for
"I don't know" (i.e., when midr.txt is absent on arm64), and then update
the skipping logic to explicitly log that this was the case and that it
will be pessimistically assuming that we are on a Cavium. This ensures
that we don't unintentionally run these tests on a Cavium when
ZX_DEBUG_ASSERT_IMPLEMENTED is zero. Not ideal, but we should still have
sufficient test coverage for these cases on other arm64 hardware for a
variety of configurations.

Original-Fixed: 467750183
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1451512
Original-Revision: bc8908aed17d38368e3825d87bc9c5043d060bf7
GitOrigin-RevId: ef3559d7d075d636657ba37e9975efee4b375a5b
Change-Id: I99224c75f048c4bcfda294ecb9e8f90f784eb50b