[roll] Roll fuchsia [sdk] Add strict audit script.
This is a vibe-coded tool that identifies all the `strict` items in the
SDK surface, and tags them with the `strict_audit` attribute, indicating
that they should be audited for evolvability. See
https://fxbug.dev/42055534 for more details.

This is probably a one-time-use script, but I dunno, it feels better to
check it in than to not. The next CL in the chain will actually run the
thing and add the annotations.

Note that this also adds the "locations" to fields in the platform IR.
It brings the resulting IR file from like ~20MB to ~30MB, but I don't
think that's a big deal.

Original-Bug: 42055534

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1444509
Original-Revision: 47f0b509fa37d70d68f5b069219df3254d89e778
GitOrigin-RevId: c201c924562fbc7a2ad7ad57d96730283ab734ea
Change-Id: I5b6905f71d9b92c46d6e1094a6075395180c4319