[roll] Roll fuchsia [f2fs_reader] Properly handle checkpoints.
This change addresses a bug where the f2fs reader could fail if it
selected a checkpoint that appeared valid but was associated with a
corrupt or unreadable summary block.

Previously, the reader would select the newest checkpoint based on its
version number alone, without fully validating the data it pointed to.
(In particular, the problem I noticed we failed to identify was an
invalid compact summary block).

This change makes the reader more robust by treating the checkpoint and
its summary block as a single unit for validation. It now attempts to
initialize the filesystem from each available checkpoint, from newest to
oldest, and uses the first one that is fully and correctly loaded. This
ensures that a bad checkpoint pack does not prevent the reader from
starting if a good one is available.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1455529
Original-Revision: cfa73670ab94dec76ca06676e1198fe4c7432dd3
GitOrigin-RevId: df289e3027dda807215fac04d4146e1cac347f9f
Change-Id: I9d8e5f8997f53825070620871474cf4e9ad7a2dd