[roll] Roll fuchsia [kernel][ktl] Soft transition for libc++ std::align change
This is the first half of a soft transition for ktl to work with
the libc++ header implementation of std::align -> ktl::align.

This changes the out-of-line definition to use manual C++ name
mangling as a means to hide the definition's true name from the
compiler.  This doesn't change what gets defined at all.

Once the new libc++ rolls in, its header will have an inline
definition for std::align, so that definition will not be used
any more and can be removed.

Original-Bug: 470118858
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1456168
Original-Revision: 6295c13e606f1b9eaeceff625b65ea0b6c39b27d
GitOrigin-RevId: 4e33b6afa0c4b145e3e471af25f70d68d30a3501
Change-Id: I1360e5362a8738d7e0b44af8bc4334b9e7807002