[roll] Roll fuchsia [ffx] Shorten the description of the options and commands
This cleans up the descriptions of main commands and options so that
they fit in a single line in the terminal.

Thus, when the user types `ffx` the visual output is much cleaner and
easier to read. In most cases, especially for the description of the
options, it is more concise and more obvious on what it does.

Resulting CLI golden files are also updated.

Before: https://paste.googleplex.com/4985969088528384?raw
After: https://paste.googleplex.com/5801500217573376?raw

Test: Run `ffx`
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1458173
Original-Revision: 996e882430239372a7e3e9386ef4724371f81875
GitOrigin-RevId: 1ddcb9b481449e26fd1c5bd1b9486c14db54ac21
Change-Id: I465924efc9de432fa0c3f988426725eb785e83a8