[roll] Roll fuchsia [kernel][usercopy] copy_array_* remove duplication
The copy_array_* functions in usercopy have a variants that either copy
from the start of the passed array, or from a provided offset. The code
is almost identical, so duplication can be removed.

This is applied to all 4 of the copy_array_* functions, copy to and from
user and with or without capture faults.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1448848
Original-Revision: 3c674543d2e44e01ff174ec6c97fb927f5f978d9
GitOrigin-RevId: 3cc4f65c079fdd94b0071503ee760e6a4d089409
Change-Id: Ib76d0c4fd688458fc57d36365e4a2599c478846a