[roll] Roll fuchsia [bazel][boringssl] Reduce copy span of @boringssl repository.
For historical reasons, the @boringssl repository must currently be
populated by hard-linking the content of //third_party/boringssl
into the repository directory, and creating a patched BUILD.bazel
file for it. That source directory is huge though.

This CL reduces the amount of files being copied, from 569 MiB
to 139 MiB, by omitting a few sub-directories that are not needed
by the build.

It also improves the repository rule's tracking of input files a
bit, ensuring it will be run when all its inputs are modified
(the previous version ignored modifications to the hardlink script
and the content hash file, in particular!).

This is only a temporary work-around until a better solution, such
as https://fxrev.dev/1442608, which completely gets rid of the
copy, can be submitted to the tree.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1458728
Original-Revision: a9bac74aca2e6b7985b8ded999c94061b13bd686
GitOrigin-RevId: e4e0bf3c2ef25a71c27c761d1f51c67dba0b140a
Change-Id: I54937852b0bb6d35af0139309bd24a1e05135cc4